Move `Router` component and dependencies to `forms-common`
This component is shared between the groups and login forms.
